// src/lib/types.ts
export interface AboutData {
    hero: {
        title: string
        description: string
    }
    story: {
        title: string
        content: string
    }
    mission: {
        title: string
        description: string
    }
    vision: {
        title: string
        description: string
    }
    values: {
        title: string
        description: string
    }[]
}